Fixed a problem with the database involving tinyints, changed notes to text data type. Tables should all work.

+++ AssignmentHub/db/migrate/20120223013404_create_users.rb
@@ -7,7 +7,7 @@ class CreateUsers < ActiveRecord::Migration
       t.string :last
       t.string :email
       t.string :phone
-      t.tinyint :user_type
+      t.integer :user_type
 
       t.timestamps
     end
